+++ class/usernavhistory.class.php
@@ -549,6 +549,11 @@ class UserNavHistory extends CommonObject
 		elseif ($elementtype == 'contact')  {
 			$module = 'societe';
 		}
+		elseif ($elementtype == 'category')  {
+			$classpath = 'categories/class';
+			$module ='categorie';
+			$myobject='categorie';
+		}
 		elseif ($elementtype == 'propal')  {
 			$classpath = 'comm/propal/class';
 		}
@@ -699,6 +704,12 @@ class UserNavHistory extends CommonObject
 			$classfile='adherent_type';
 			$classname='AdherentType';
 		}
+		elseif ($elementtype == 'facturerec') {
+                        $classfile = 'facture-rec';
+                        $classpath = 'compta/facture/class';
+                        $module='facture';
+                        $myobject='FactureRec';
+                }
 // @todo : Sondage
 //var_dump($conf);
 		if (!empty($conf->$module->enabled))