+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## 1.0
 
+- FIX: Compatibility V16 - Family - *24/06/2022* - 1.0.6
+- FIX : object category was not compatible *17/07/2022* - 1.0.5
+- FIX : object facturerec was not compatible *23/06/2022* - 1.0.4
+- FIX : do not display usernavhistory in print mode *23/06/2022* - 1.0.3
 - FIX : object picto max size *14/06/2022* - 1.0.2
 - FIX : Do not display deleted elements *05/05/2022* - 1.0.1
 Initial version